# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TotalReclaw Server",
        "description": "\n    Zero-knowledge encrypted memory vault API.\n\n    ## Authentication\n\n    All endpoints (except /register and /health) require authentication using\n    a Bearer token in the Authorization header:\n\n    ```\n    Authorization: Bearer <hex-encoded-auth-key>\n    ```\n\n    The auth_key is derived from the user's master password using HKDF-SHA256:\n    ```\n    auth_key = HKDF-SHA256(master_password, salt, \"totalreclaw-auth-v1\")\n    ```\n\n    ## Zero-Knowledge Design\n\n    - Server NEVER sees the master password\n    - Server NEVER sees the encryption key\n    - Server NEVER sees plaintext memories\n    - Server only stores encrypted blobs and blind indices\n    ",
        "version": "0.3.1"
    },
    "paths": {
...
        "/v1/search": {
            "post": {
                "tags": [
                    "search"
                ],
                "summary": "Search",
                "description": "Search for facts using blind indices.\n\nThe client sends:\n- user_id: User's UUID (must match authenticated user)\n- trapdoors: List of blind trapdoors (SHA-256 hashes of LSH buckets + keywords)\n- max_candidates: Maximum number of candidates to return\n- min_decay_score: Filter by minimum decay score\n\nThe server:\n1. Validates user_id matches authenticated user\n2. Queries facts where blind_indices overlap with trapdoors\n3. Returns encrypted blobs for client-side re-ranking\n\nThe server NEVER sees the plaintext or knows what the search terms are.\nThe blind indices are one-way SHA-256 hashes.",
                "operationId": "search_v1_search_post",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchRequestJSON"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchResponseJSON"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/facts/{fact_id}": {
            "delete": {
                "tags": [
                    "search"
                ],
                "summary": "Delete Fact",
                "description": "Soft delete a fact (create tombstone).\n\nThe fact is marked as inactive and a tombstone is created.\nTombstones are retained for 30 days for undo capability.",
                "operationId": "delete_fact_v1_facts__fact_id__delete",
                "parameters": [
                    {
                        "name": "fact_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Fact Id"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResponseJSON"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/export": {
            "get": {
                "tags": [
                    "search"
                ],
                "summary": "Export Facts",
                "description": "Export user facts with cursor-based pagination.\n\nReturns active facts for the authenticated user, paginated to prevent\nOOM on large accounts. Use the returned `cursor` value as the `cursor`\nquery parameter in the next request to fetch the next page.\n\nQuery parameters:\n- limit: Number of facts per page (default: 1000, max: 5000)\n- cursor: Last fact_id from previous page (omit for first page)",
                "operationId": "export_facts_v1_export_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 5000,
                            "minimum": 1,
                            "description": "Number of facts per page",
                            "default": 1000,
                            "title": "Limit"
                        },
                        "description": "Number of facts per page"
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Last fact_id from previous page",
                            "title": "Cursor"
                        },
                        "description": "Last fact_id from previous page"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExportResponseJSON"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeleteResponseJSON": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "error_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Code"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    }
                },
                "type": "object",
                "required": [
                    "success"
                ],
                "title": "DeleteResponseJSON",
                "description": "JSON delete response."
            },
            "ExportResponseJSON": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "error_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Code"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    },
                    "facts": {
                        "anyOf": [
                            {
                                "items": {
                                    "additionalProperties": true,
                                    "type": "object"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Facts"
                    },
                    "cursor": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cursor"
                    },
                    "has_more": {
                        "type": "boolean",
                        "title": "Has More",
                        "default": false
                    },
                    "total_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Count"
                    }
                },
                "type": "object",
                "required": [
                    "success"
                ],
                "title": "ExportResponseJSON",
                "description": "JSON export response with cursor-based pagination."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SearchRequestJSON": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "maxLength": 100,
                        "title": "User Id",
                        "description": "User's UUID"
                    },
                    "trapdoors": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "maxItems": 1000,
                        "title": "Trapdoors",
                        "description": "Blind trapdoors (SHA-256 hashes)"
                    },
                    "max_candidates": {
                        "type": "integer",
                        "maximum": 10000.0,
                        "minimum": 1.0,
                        "title": "Max Candidates",
                        "description": "Maximum candidates to return",
                        "default": 3000
                    },
                    "min_decay_score": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Min Decay Score",
                        "description": "Minimum decay score filter",
                        "default": 0.0
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "trapdoors"
                ],
                "title": "SearchRequestJSON",
                "description": "JSON search request."
            },
            "SearchResponseJSON": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "error_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Code"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    },
                    "results": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/SearchResultJSON"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Results"
                    },
                    "total_candidates": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Candidates"
                    }
                },
                "type": "object",
                "required": [
                    "success"
                ],
                "title": "SearchResponseJSON",
                "description": "JSON search response."
            },
            "SearchResultJSON": {
                "properties": {
                    "fact_id": {
                        "type": "string",
                        "title": "Fact Id"
                    },
                    "encrypted_blob": {
                        "type": "string",
                        "title": "Encrypted Blob"
                    },
                    "decay_score": {
                        "type": "number",
                        "title": "Decay Score"
                    },
                    "timestamp": {
                        "type": "integer",
                        "title": "Timestamp"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    }
                },
                "type": "object",
                "required": [
                    "fact_id",
                    "encrypted_blob",
                    "decay_score",
                    "timestamp",
                    "version"
                ],
                "title": "SearchResultJSON",
                "description": "JSON search result."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}